@@ -7,7 +7,7 @@ description: A flutter appliction listing all the widgets covered in Flutter wid
 # Both the version and the builder number may be overridden in flutter
 # build by specifying --build-name and --build-number, respectively.
 # Read more about versioning at semver.org.
-version: 1.0.5+6
+version: 1.0.14+16
 
 dependencies:
   flutter:
@@ -19,13 +19,16 @@ dependencies:
   flutter_svg: ^0.14.0
   url_launcher: ^5.1.2
   http: ^0.12.0+1
-  youtube_player: ^3.5.0
+  youtube_player_flutter: ^6.1.1
   transparent_image: ^1.0.0
   package_info: ^0.4.0+6
-  firebase_core: ^0.4.0+1
-  firebase_analytics: ^4.0.2
-  firebase_remote_config: ^0.2.0+5
-  webview_flutter: ^0.3.11+2
+  firebase_core: ^0.4.2
+  firebase_analytics: ^5.0.9
+  firebase_remote_config: ^0.3.0
+  shared_preferences: ^0.5.3+4
+  firebase_messaging: ^6.0.9
+  firebase_admob: ^0.9.3+2
+  provider: ^4.0.1
 
 dev_dependencies:
   flutter_test:
@@ -49,6 +52,7 @@ flutter:
       - assets/images/github.png
       - assets/images/linkedin.png
       - assets/images/twitter.png
+      - assets/images/slack_icon.svg
 
   fonts:
       - family: Ubuntu
