@@ -1,11 +1,12 @@
+import 'dart:async';
+
 import 'package:flutter_widget_guide/model/demo_model.dart';
 import 'package:http/http.dart' as http;
-import 'dart:async';
 
 String url = 'https://jsonplaceholder.typicode.com/posts';
 
 ///Method for GET Request
-Future<Demo> getDemoResponse() async{
+Future<Demo> getDemoResponse() async {
   final response = await http.get('$url/1');
   return responseFromJson(response.body);
 }
