@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_widget_guide/utils.dart';
 
+import '../Ads.dart';
 import '../Code.dart';
 import '../CodeScreen.dart';
 
@@ -11,9 +12,16 @@ class AbsorbPointerWidget extends StatefulWidget {
 
 class _AbsorbPointerWidgetState extends State<AbsorbPointerWidget> {
   static final GlobalKey<ScaffoldState> _scaffoldKey =
-  GlobalKey<ScaffoldState>();
+      GlobalKey<ScaffoldState>();
   var _actionIcon = Icons.blur_on;
   var _isAbsorbing = false;
+
+  @override
+  void dispose() {
+    // Hide banner ad if it isn't already
+    Ads.hideBannerAd();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -32,21 +40,21 @@ class _AbsorbPointerWidgetState extends State<AbsorbPointerWidget> {
           IconButton(
             icon: Icon(_actionIcon),
             onPressed: () => setState(() {
-                  _isAbsorbing == false
-                      ? _isAbsorbing = true
-                      : _isAbsorbing = false;
+              _isAbsorbing == false
+                  ? _isAbsorbing = true
+                  : _isAbsorbing = false;
 
-                  _isAbsorbing
-                      ? _actionIcon = Icons.blur_off
-                      : _actionIcon = Icons.blur_on;
+              _isAbsorbing
+                  ? _actionIcon = Icons.blur_off
+                  : _actionIcon = Icons.blur_on;
 
-                  _scaffoldKey.currentState.showSnackBar(
-                    SnackBar(
-                      content: Text("Absorbing Property changed to $_isAbsorbing"),
-                      duration: Duration(milliseconds: 1000),
-                    ),
-                  );
-                }),
+              _scaffoldKey.currentState.showSnackBar(
+                SnackBar(
+                  content: Text("Absorbing Property changed to $_isAbsorbing"),
+                  duration: Duration(milliseconds: 1000),
+                ),
+              );
+            }),
           ),
           IconButton(
             icon: Icon(Icons.code),
@@ -76,7 +84,7 @@ class _AbsorbPointerWidgetState extends State<AbsorbPointerWidget> {
                 child: Text(
                   "Absorb Pointer Enabled: $_isAbsorbing",
                   style: TextStyle(
-                      color: Colors.black87,
+                      color: Theme.of(context).backgroundColor,
                       fontSize: 14.0,
                       fontFamily: Utils.ubuntuRegularFont),
                   textAlign: TextAlign.center,
