@@ -3,6 +3,7 @@ import 'dart:math' as math;
 import 'package:flutter/material.dart';
 import 'package:flutter_widget_guide/utils.dart';
 
+import '../Ads.dart';
 import '../Code.dart';
 import '../CodeScreen.dart';
 
@@ -22,6 +23,8 @@ class _AnimatedBuilderWidgetState extends State<AnimatedBuilderWidget>
   @override
   void initState() {
     super.initState();
+
+    Ads.hideBannerAd();
 
     _controller1 = AnimationController(
       duration: const Duration(seconds: 10),
@@ -73,12 +76,12 @@ class _AnimatedBuilderWidgetState extends State<AnimatedBuilderWidget>
           IconButton(
             icon: Icon(Icons.code),
             onPressed: () => Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) =>
-                        CodeScreen(code: Code.animatedBuilderCode),
-                  ),
-                ),
+              context,
+              MaterialPageRoute(
+                builder: (context) =>
+                    CodeScreen(code: Code.animatedBuilderCode),
+              ),
+            ),
           )
         ],
       ),
@@ -164,11 +167,11 @@ class _AnimatedBuilderWidgetState extends State<AnimatedBuilderWidget>
                   animation: widthAnimation,
                   child: null,
                   builder: (context, child) => Container(
-                        height: 100,
-                        width: widthAnimation.value,
-                        color: Colors.indigo,
-                        child: child,
-                      ),
+                    height: 100,
+                    width: widthAnimation.value,
+                    color: Colors.indigo,
+                    child: child,
+                  ),
                 ),
               ),
             ],
@@ -179,10 +182,10 @@ class _AnimatedBuilderWidgetState extends State<AnimatedBuilderWidget>
       floatingActionButton: FloatingActionButton.extended(
         backgroundColor: Colors.green,
         onPressed: () => setState(() {
-              _controller1.repeat();
-              _controller2.forward();
-              _controller3.forward();
-            }),
+          _controller1.repeat();
+          _controller2.forward();
+          _controller3.forward();
+        }),
         icon: Icon(
           Icons.movie_filter,
           color: Colors.white,
