@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_widget_guide/utils.dart';
 
+import '../Ads.dart';
 import '../Code.dart';
 import '../CodeScreen.dart';
 
@@ -18,11 +19,15 @@ class _AspectRatioWidgetState extends State<AspectRatioWidget> {
   @override
   void initState() {
     super.initState();
+
     /// Fix portrait mode
     SystemChrome.setPreferredOrientations([
       DeviceOrientation.portraitUp,
       DeviceOrientation.portraitDown,
     ]);
+
+    //Hide banner ad if it isn't already hidden
+    Ads.hideBannerAd();
   }
 
   @override
@@ -53,6 +58,7 @@ class _AspectRatioWidgetState extends State<AspectRatioWidget> {
           children: <Widget>[
             Padding(
               padding: const EdgeInsets.all(8.0),
+
               /// Aspect Ratio widget sizes itself freely
               /// depending on available space
               child: AspectRatio(
@@ -67,10 +73,10 @@ class _AspectRatioWidgetState extends State<AspectRatioWidget> {
               margin: EdgeInsets.only(left: 12, right: 12),
               child: Wrap(
                 children: <Widget>[
+                  _OptionItem(aspectRatio1, aspectRatio, _onRatioChanged,
+                      'Ratio: 16/9'),
                   _OptionItem(
-                      aspectRatio1, aspectRatio, _onRatioChanged, 'Ratio: 16/9'),
-                  _OptionItem(aspectRatio2, aspectRatio, _onRatioChanged,
-                      'Ratio: 3/2'),
+                      aspectRatio2, aspectRatio, _onRatioChanged, 'Ratio: 3/2'),
                 ],
               ),
             )
