@@ -3,6 +3,7 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:flutter_widget_guide/utils.dart';
 
+import '../Ads.dart';
 import '../Code.dart';
 import '../CodeScreen.dart';
 
@@ -13,6 +14,14 @@ class BackdropFilterWidget extends StatefulWidget {
 
 class _BackdropFilterWidgetState extends State<BackdropFilterWidget> {
   var _blurValue = 0.0;
+
+  @override
+  void initState() {
+    //Hide banner ad if it isn't already hidden
+    Ads.hideBannerAd();
+    super.initState();
+  }
+
 
   @override
   Widget build(BuildContext context) {
@@ -30,12 +39,11 @@ class _BackdropFilterWidgetState extends State<BackdropFilterWidget> {
           IconButton(
             icon: Icon(Icons.code),
             onPressed: () => Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) =>
-                        CodeScreen(code: Code.backDropFilterCode),
-                  ),
-                ),
+              context,
+              MaterialPageRoute(
+                builder: (context) => CodeScreen(code: Code.backDropFilterCode),
+              ),
+            ),
           )
         ],
       ),
