@@ -4,6 +4,8 @@ import 'package:flutter_widget_guide/Code.dart';
 import 'package:flutter_widget_guide/CodeScreen.dart';
 import 'package:flutter_widget_guide/utils.dart';
 
+import '../Ads.dart';
+
 class ExpandedWidget extends StatefulWidget {
   @override
   _ExpandedWidgetState createState() => _ExpandedWidgetState();
@@ -12,10 +14,19 @@ class ExpandedWidget extends StatefulWidget {
 class _ExpandedWidgetState extends State<ExpandedWidget> {
   ///Flex factor for 1st widget
   int _flexFactor0 = 1;
+
   ///Flex factor for 2nd widget
   int _flexFactor1 = 2;
+
   ///Flex factor for 3rd widget
   int _flexFactor2 = 1;
+
+  @override
+  void initState() {
+    //Hide banner ad if it isn't already hidden
+    Ads.hideBannerAd();
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -33,49 +44,49 @@ class _ExpandedWidgetState extends State<ExpandedWidget> {
           IconButton(
             icon: Icon(
               Icons.info_outline,
-              color: Color(0xFF323232),
+              color: Theme.of(context).backgroundColor,
             ),
             onPressed: () => showDialog(
-                  context: context,
-                  builder: (_) => AlertDialog(
-                        contentPadding: EdgeInsets.all(0),
-                        titlePadding: EdgeInsets.only(top: 28),
-                        title: Center(
-                          child: Column(
-                            children: <Widget>[
-                              Text(
-                                "Description",
-                                style: TextStyle(
-                                    fontSize: 20.0,
-                                    fontWeight: FontWeight.bold,
-                                    fontFamily: Utils.ubuntuRegularFont),
-                              ),
-                              FlatButton(
-                                onPressed: () =>
-                                    Utils.launchURL(Utils.expandedSourceURL),
-                                splashColor: Colors.transparent,
-                                highlightColor: Colors.transparent,
-                                // makes highlight invisible too
-                                child: Text('(Source Code)',
-                                    style: TextStyle(
-                                        fontSize: 14.0,
-                                        fontWeight: FontWeight.w300,
-                                        fontFamily: Utils.crimsonRegularFont,
-                                        color: Colors.cyan)),
-                              ),
-                            ],
-                          ),
-                        ),
-                        content: Container(
-                          padding: EdgeInsets.all(0),
-                          margin: EdgeInsets.all(0),
-                          child: SvgPicture.asset(
-                            Utils.expanded_img,
-                            semanticsLabel: "Expanded Description",
-                          ),
-                        ),
+              context: context,
+              builder: (_) => AlertDialog(
+                contentPadding: EdgeInsets.all(0),
+                titlePadding: EdgeInsets.only(top: 28),
+                title: Center(
+                  child: Column(
+                    children: <Widget>[
+                      Text(
+                        "Description",
+                        style: TextStyle(
+                            fontSize: 20.0,
+                            fontWeight: FontWeight.bold,
+                            fontFamily: Utils.ubuntuRegularFont),
                       ),
+                      FlatButton(
+                        onPressed: () =>
+                            Utils.launchURL(Utils.expandedSourceURL),
+                        splashColor: Colors.transparent,
+                        highlightColor: Colors.transparent,
+                        // makes highlight invisible too
+                        child: Text('(Source Code)',
+                            style: TextStyle(
+                                fontSize: 14.0,
+                                fontWeight: FontWeight.w300,
+                                fontFamily: Utils.crimsonRegularFont,
+                                color: Colors.cyan)),
+                      ),
+                    ],
+                  ),
                 ),
+                content: Container(
+                  padding: EdgeInsets.all(0),
+                  margin: EdgeInsets.all(0),
+                  child: SvgPicture.asset(
+                    Utils.expanded_img,
+                    semanticsLabel: "Expanded Description",
+                  ),
+                ),
+              ),
+            ),
           ),
           IconButton(
             icon: Icon(Icons.code),
@@ -114,11 +125,12 @@ class _ExpandedWidgetState extends State<ExpandedWidget> {
                           height: 48.0,
                         ),
                         onTap: () => setState(() {
-                              _flexFactor0++;
-                            }),
+                          _flexFactor0++;
+                        }),
                       ),
                     ),
                   ),
+
                   ///Second flexible widget
                   Expanded(
                     flex: _flexFactor1,
@@ -138,11 +150,12 @@ class _ExpandedWidgetState extends State<ExpandedWidget> {
                           height: 48.0,
                         ),
                         onTap: () => setState(() {
-                              _flexFactor1++;
-                            }),
+                          _flexFactor1++;
+                        }),
                       ),
                     ),
                   ),
+
                   ///Third flexible widget
                   Expanded(
                     flex: _flexFactor2,
@@ -162,14 +175,15 @@ class _ExpandedWidgetState extends State<ExpandedWidget> {
                           height: 48.0,
                         ),
                         onTap: () => setState(() {
-                              _flexFactor2++;
-                            }),
+                          _flexFactor2++;
+                        }),
                       ),
                     ),
                   ),
                 ],
               ),
             ),
+
             ///Click on this button to RESET the values of flex values
             Container(
               margin: EdgeInsets.all(16.0),
@@ -177,10 +191,10 @@ class _ExpandedWidgetState extends State<ExpandedWidget> {
                 textColor: Colors.white,
                 color: Colors.red,
                 onPressed: () => setState(() {
-                      _flexFactor0 = 1;
-                      _flexFactor1 = 2;
-                      _flexFactor2 = 1;
-                    }),
+                  _flexFactor0 = 1;
+                  _flexFactor1 = 2;
+                  _flexFactor2 = 1;
+                }),
                 child: Text('RESET'),
               ),
             ),
