@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_widget_guide/utils.dart';
 
+import '../Ads.dart';
 import '../Code.dart';
 import '../CodeScreen.dart';
 
@@ -45,6 +46,9 @@ class _FadeTransitionWidgetState extends State<FadeTransitionWidget>
       else if (status == AnimationStatus.dismissed) _controller.forward();
     });
 
+    //Hide banner ad if it isn't already hidden
+    Ads.hideBannerAd();
+
     super.initState();
   }
 
@@ -71,11 +75,11 @@ class _FadeTransitionWidgetState extends State<FadeTransitionWidget>
           IconButton(
             icon: Icon(Icons.code),
             onPressed: () => Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => CodeScreen(code: Code.fadeTransitionCode),
-                  ),
-                ),
+              context,
+              MaterialPageRoute(
+                builder: (context) => CodeScreen(code: Code.fadeTransitionCode),
+              ),
+            ),
           )
         ],
       ),
@@ -94,9 +98,9 @@ class _FadeTransitionWidgetState extends State<FadeTransitionWidget>
       floatingActionButton: FloatingActionButton.extended(
         backgroundColor: Colors.lightBlue,
         onPressed: () => setState(() {
-              ///We start the animation by calling this
-              _controller.forward();
-            }),
+          ///We start the animation by calling this
+          _controller.forward();
+        }),
         icon: Icon(
           Icons.movie_filter,
           color: Colors.white,
