@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_widget_guide/utils.dart';
 
+import '../Ads.dart';
 import '../Code.dart';
 import '../CodeScreen.dart';
 
@@ -15,17 +16,25 @@ class _FittedBoxWidgetState extends State<FittedBoxWidget> {
   var _heightValue = 400.0;
 
   @override
+  void initState() {
+    //Hide banner ad if it isn't already hidden
+    Ads.hideBannerAd();
+    super.initState();
+  }
+
+
+  @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         centerTitle: true,
         title: Text(
-              'FittedBox Widget',
-              style: TextStyle(
-                  fontSize: 20.0,
-                  fontWeight: FontWeight.bold,
-                  fontFamily: Utils.ubuntuRegularFont),
-            ),
+          'FittedBox Widget',
+          style: TextStyle(
+              fontSize: 20.0,
+              fontWeight: FontWeight.bold,
+              fontFamily: Utils.ubuntuRegularFont),
+        ),
         actions: <Widget>[
           IconButton(
             icon: Icon(Icons.code),
@@ -77,7 +86,7 @@ class _FittedBoxWidgetState extends State<FittedBoxWidget> {
                 padding: EdgeInsets.only(left: 12, right: 12, bottom: 12),
                 child: Text(
                   'Slide the slider widget to change the value of height\n'
-                      'to observe how the widget adjusts itself',
+                  'to observe how the widget adjusts itself',
                   style: TextStyle(
                       color: Colors.grey[400],
                       fontSize: 14.0,
