@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_widget_guide/utils.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_widget_guide/utils.dart';
 
+import '../Ads.dart';
 import '../Code.dart';
 import '../CodeScreen.dart';
 
@@ -11,7 +12,6 @@ class FlexibleWidget extends StatefulWidget {
 }
 
 class _FlexibleWidgetState extends State<FlexibleWidget> {
-
   @override
   void initState() {
     super.initState();
@@ -19,6 +19,9 @@ class _FlexibleWidgetState extends State<FlexibleWidget> {
       DeviceOrientation.portraitUp,
       DeviceOrientation.portraitDown,
     ]);
+
+    //Hide banner ad if it isn't already hidden
+    Ads.hideBannerAd();
   }
 
   @override
@@ -142,7 +145,8 @@ class _FlexibleWidgetState extends State<FlexibleWidget> {
                             fontSize: 28.0),
                       ),
                       Padding(
-                        padding: const EdgeInsets.all(8.0),
+                        padding: const EdgeInsets.only(
+                            top: 8.0, right: 8.0, left: 8.0),
                         child: Text(
                           "(1 Flex)",
                           style: TextStyle(
