@@ -1,10 +1,15 @@
+import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:firebase_remote_config/firebase_remote_config.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter_svg/svg.dart';
 import 'package:flutter_widget_guide/bloc/list_bloc.dart';
 import 'package:flutter_widget_guide/model/list_Item.dart';
 import 'package:flutter_widget_guide/utils.dart';
 import 'package:flutter_widget_guide/widgets/home_list_item.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
+import '../Ads.dart';
 import '../profile_screen.dart';
 import 'WebViewWidget.dart';
 
@@ -13,19 +18,129 @@ class HomePage extends StatefulWidget {
   _HomePageState createState() => _HomePageState();
 }
 
-class _HomePageState extends State<HomePage> {
+/// WidgetsBindingObserver helps to keep track of the app lifecycle state
+class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
   BuildContext _buildContext;
   var versionNumber;
-  String appLink = "https://play.google.com/store/apps/details?id=com.annsh.flutterwidgetguide";
+  bool isFabVisible = true;
+  bool hasJoinedSlack = false;
+  ScrollController _hideButtonController;
+  bool isCheckBoxChecked = false;
+  String appLink =
+      "https://play.google.com/store/apps/details?id=com.annsh.flutterwidgetguide";
   final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
+  FirebaseMessaging _fcm;
+
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) {
+    //print("The state is $state");
+    /// do something here to release FCM instance so that it doesn't configure
+    /// again when user comes back to the app from Recent Apps
+    /// Nothing can be done as of now.
+    /// Link to the issue:
+    /// https://github.com/FirebaseExtended/flutterfire/issues/1060
+    super.didChangeAppLifecycleState(state);
+  }
+
+  @override
+  void dispose() {
+    WidgetsBinding.instance.removeObserver(this);
+    //Dispose the Ad if it isn't already
+    Ads.hideBannerAd();
+    super.dispose();
+  }
 
   @override
   void initState() {
     super.initState();
+    WidgetsBinding.instance.addObserver(this);
+    //Initialize Firebase Admob
+    Ads.initialize();
+    _fcm = FirebaseMessaging();
     Utils.getVersion().then((value) {
       versionNumber = value;
     });
     setupRemoteConfig();
+    _getValueFromSP();
+    isFabVisible = true;
+    _hideButtonController = new ScrollController();
+    _hideButtonController.addListener(
+      () {
+        if (_hideButtonController.position.userScrollDirection ==
+            ScrollDirection.reverse) {
+          if (isFabVisible == true) {
+            setState(() => isFabVisible = false);
+          }
+        } else {
+          if (_hideButtonController.position.userScrollDirection ==
+              ScrollDirection.forward) {
+            if (isFabVisible == false) {
+              setState(() => isFabVisible = true);
+            }
+          }
+        }
+      },
+    );
+
+    /// To check if the user is already subscribed to the topic
+    _getIsSubscribed().then((isSubscribed) {
+      if (!isSubscribed) {
+        _fcm.subscribeToTopic("learn").then((value) {
+          _setIsSubscribed(true);
+        }, onError: (e) {
+          _setIsSubscribed(false);
+        });
+      }
+    });
+
+    //_getIsFcmConfigured().then((value) {
+    // if (!value) {
+    _fcm.configure(
+      onMessage: (Map<String, dynamic> message) async {
+        /// Called whenever the app is in foreground and receives a notification
+        /// A dialog box is shown to the user in this case
+        showDialog(
+          context: context,
+          builder: (context) => AlertDialog(
+            content: ListTile(
+              title: Text("A new message from the developer!"),
+              subtitle: Padding(
+                padding: const EdgeInsets.only(top: 8.0),
+                child: Text("Click open to visit the link"),
+              ),
+            ),
+            actions: <Widget>[
+              FlatButton(
+                child: Text(
+                  'Cancel',
+                  style: TextStyle(color: Colors.grey[500]),
+                ),
+                onPressed: () => Navigator.of(context).pop(),
+              ),
+              FlatButton(
+                child: Text('Open'),
+                onPressed: () => {
+                  Navigator.of(context).pop(),
+                  _takeNotificationAction(message, context, true),
+                },
+              ),
+            ],
+          ),
+        );
+      },
+      onBackgroundMessage: backgroundHandle,
+      onLaunch: (Map<String, dynamic> message) async {
+        /// Called whenever the app is killed and receives a notification
+        _takeNotificationAction(message, context, false);
+      },
+      onResume: (Map<String, dynamic> message) async {
+        /// Called whenever the app is running in background
+        /// and receives a notification
+        _takeNotificationAction(message, context, false);
+      },
+    );
+    // }
+    //});
   }
 
   @override
@@ -38,9 +153,107 @@ class _HomePageState extends State<HomePage> {
         data: Theme.of(context).copyWith(
           canvasColor: Colors.transparent,
         ),
-        child: Scaffold(
-          key: _scaffoldKey,
-          body: sliverWidgetList(),
+        child: WillPopScope(
+          child: Scaffold(
+            key: _scaffoldKey,
+            body: sliverWidgetList(),
+            floatingActionButtonLocation:
+                FloatingActionButtonLocation.centerFloat,
+            floatingActionButton: Visibility(
+              visible: isFabVisible && !hasJoinedSlack,
+              child: FloatingActionButton.extended(
+                backgroundColor: Color(0xFFffffff),
+                icon: Container(
+                  height: 24,
+                  width: 24,
+                  child: SvgPicture.asset(
+                    Utils.slack_img,
+                    semanticsLabel: "Join Slack",
+                  ),
+                ),
+                label: Text(
+                  "Join us",
+                  style: TextStyle(
+                      color: Colors.black87,
+                      fontFamily: Utils.ubuntuRegularFont),
+                ),
+                onPressed: () => {
+                  showDialog(
+                    context: context,
+
+                    /// StatefulBuilder is used here to make setState work on AlertDialog
+                    /// For checkbox state functionality
+                    builder: (context) => StatefulBuilder(
+                      builder: (context, setState) {
+                        return AlertDialog(
+                          title: Center(
+                            child: Text(
+                              "Join us at\nFlutter Worldwide",
+                              textAlign: TextAlign.center,
+                              style: TextStyle(
+                                  fontSize: 20.0,
+                                  fontWeight: FontWeight.bold,
+                                  fontFamily: Utils.ubuntuRegularFont),
+                            ),
+                          ),
+                          content: Column(
+                            mainAxisSize: MainAxisSize.min,
+                            crossAxisAlignment: CrossAxisAlignment.center,
+                            children: <Widget>[
+                              Text(
+                                "My main motive here is to create a community of flutter developers"
+                                " from all around the world. Join us to expand your knowledge on Flutter"
+                                " with the rest of the world.",
+                                style: TextStyle(
+                                  fontSize: 14.0,
+                                  fontFamily: Utils.ubuntuRegularFont,
+                                ),
+                                textAlign: TextAlign.center,
+                              ),
+                              Padding(
+                                padding: const EdgeInsets.all(8.0),
+                                child: OutlineButton(
+                                    child: Text("Get an Invite"),
+                                    onPressed: () => Utils.launchURL(
+                                        "${Utils.slack_invite}"),
+                                    borderSide: BorderSide(color: Colors.blue),
+                                    shape: RoundedRectangleBorder(
+                                        borderRadius:
+                                            BorderRadius.circular(30.0))),
+                              ),
+                              Padding(
+                                padding: const EdgeInsets.all(8.0),
+                                child: Row(
+                                  mainAxisSize: MainAxisSize.max,
+                                  mainAxisAlignment: MainAxisAlignment.center,
+                                  children: <Widget>[
+                                    Text("Already joined? \nCheck to hide FAB"),
+                                    Checkbox(
+                                      value: isCheckBoxChecked,
+                                      activeColor: Colors.blue,
+                                      onChanged: (bool isChecked) {
+                                        setState(
+                                          () {
+                                            isCheckBoxChecked = isChecked;
+                                          },
+                                        );
+                                        _hideFabForever(isChecked);
+                                      },
+                                    ),
+                                  ],
+                                ),
+                              ),
+                            ],
+                          ),
+                        );
+                      },
+                    ),
+                  ),
+                },
+              ),
+            ),
+          ),
+          onWillPop: _willPopCallback,
         ),
       );
 
@@ -51,6 +264,7 @@ class _HomePageState extends State<HomePage> {
         builder: (context, snapshot) {
           return snapshot.hasData
               ? CustomScrollView(
+                  controller: _hideButtonController,
                   //This is to contain Sliver Elements
                   slivers: <Widget>[
                     appBar(context),
@@ -68,7 +282,7 @@ class _HomePageState extends State<HomePage> {
   }
 
   Widget appBar(BuildContext context) => SliverAppBar(
-        backgroundColor: Colors.white,
+        backgroundColor: Theme.of(context).primaryColorDark,
         pinned: true,
         elevation: 3.0,
         forceElevated: false,
@@ -89,11 +303,11 @@ class _HomePageState extends State<HomePage> {
                     colors: Colors.cyan,
                     textColor: Colors.white,
                   ),
-                  onTap: () =>  Navigator.push(
+                  onTap: () => Navigator.push(
                     context,
                     MaterialPageRoute(
-                      builder: (context) => WebViewWidget(
-                          url: "https://flutter.dev"),
+                      builder: (context) =>
+                          WebViewWidget(url: "https://flutter.dev"),
                     ),
                   ),
                 ),
@@ -154,7 +368,6 @@ class _HomePageState extends State<HomePage> {
     }
   }
 
-
   /// Build a snackbar to notify user that a new update is available
   buildSnakbar() {
     _scaffoldKey.currentState.showSnackBar(SnackBar(
@@ -169,4 +382,153 @@ class _HomePageState extends State<HomePage> {
       ),
     ));
   }
+
+  /// Method to call when user checks checkbox
+  _hideFabForever(bool isChecked) async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    await prefs.setBool('hidefab', isChecked);
+    setState(() {
+      hasJoinedSlack = isChecked;
+    });
+  }
+
+  /// Method to get value from shared preferences
+  _getValueFromSP() async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    setState(() {
+      if (prefs.getBool('hidefab') != null) {
+        hasJoinedSlack = prefs.getBool('hidefab');
+      } else {
+        hasJoinedSlack = false;
+      }
+    });
+  }
+
+  _setIsSubscribed(bool isSubscribed) async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    prefs.setBool("subscribed", isSubscribed);
+  }
+
+  Future<bool> _getIsSubscribed() async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    if (prefs.getBool("subscribed") == null ||
+        prefs.getBool("subscribed") == false) {
+      return false;
+    } else {
+      return true;
+    }
+  }
 }
+
+/// Use this method to define some kind of a background task
+/// We are not using it, only here for learning purpose
+Future<dynamic> backgroundHandle(Map<String, dynamic> message) {
+  if (message.containsKey('data')) {
+    // Handle data message
+  }
+
+  if (message.containsKey('notification')) {
+    // Handle notification message
+  }
+
+  // Or do other work.
+}
+
+/// Method to be called whenever the app receives a notification
+_takeNotificationAction(
+    Map<String, dynamic> message, BuildContext context, bool isInside) {
+  /// If the notification was opened when app was in background/closed
+  if (!isInside) {
+    showDialog(
+      context: context,
+      builder: (_) => AlertDialog(
+        shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.all(Radius.circular(10.0))),
+        content: Builder(
+          builder: (context) {
+            return Container(
+              height: 100,
+              width: 100,
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: <Widget>[
+                  Center(
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        CircularProgressIndicator(
+                            valueColor:
+                                AlwaysStoppedAnimation<Color>(Colors.blue)),
+                        Padding(
+                          padding: const EdgeInsets.only(top: 12.0),
+                          child: Text("Loading..."),
+                        )
+                      ],
+                    ),
+                  ),
+                ],
+              ),
+            );
+          },
+        ),
+      ),
+    );
+
+    /// To display alert dialog for 3 seconds and take necessary action
+    /// after that
+    Future.delayed(
+      Duration(seconds: 3),
+      () {
+        //remove the loading dialog box
+        Navigator.of(context, rootNavigator: true).pop();
+        if (message['data']['type'] == 'update') {
+          Utils.launchURL(message['data']['url']);
+        } else {
+          Navigator.push(
+            context,
+            MaterialPageRoute(
+              builder: (context) => WebViewWidget(url: message['data']['url']),
+            ),
+          );
+        }
+      },
+    );
+  } else {
+    /// If the notification was opened from inside the app
+    if (message['data']['type'] == 'update') {
+      Utils.launchURL(message['data']['url']);
+    } else {
+      Navigator.push(
+        context,
+        MaterialPageRoute(
+          builder: (context) => WebViewWidget(url: message['data']['url']),
+        ),
+      );
+    }
+  }
+  //_setIsFcmConfigured(true);
+}
+
+Future<bool> _willPopCallback() async {
+  /// do something here to release FCM instance so that it doesn't configure
+  /// again when user comes back to the app from Recent Apps
+  /// Nothing can be done as of now.
+  /// Link to the issue:
+  /// https://github.com/FirebaseExtended/flutterfire/issues/1060
+  return true; // return true if the route to be popped
+}
+
+//_setIsFcmConfigured(bool isConfigured) async {
+//  SharedPreferences prefs = await SharedPreferences.getInstance();
+//  prefs.setBool("isConfigured", isConfigured);
+//}
+//
+//Future<bool> _getIsFcmConfigured() async {
+//  SharedPreferences prefs = await SharedPreferences.getInstance();
+//  if (prefs.getBool("isConfigured") == null ||
+//      prefs.getBool("isConfigured") == false) {
+//    return false;
+//  } else {
+//    return true;
+//  }
+//}
