@@ -11,19 +11,19 @@ class LayoutBuilderWidget extends StatelessWidget {
       appBar: AppBar(
         centerTitle: true,
         title: Text(
-            'LayoutBuilder Widget',
-            style: TextStyle(
-                fontSize: 20.0,
-                fontWeight: FontWeight.bold,
-                fontFamily: Utils.ubuntuRegularFont),
-          ),
+          'LayoutBuilder Widget',
+          style: TextStyle(
+              fontSize: 20.0,
+              fontWeight: FontWeight.bold,
+              fontFamily: Utils.ubuntuRegularFont),
+        ),
         actions: <Widget>[
           IconButton(
             icon: Icon(Icons.code),
             onPressed: () => Navigator.push(
               context,
               MaterialPageRoute(
-                builder: (context) => CodeScreen(code: Code.clipRrectCode),
+                builder: (context) => CodeScreen(code: Code.layoutBuilderCode),
               ),
             ),
           )
@@ -32,22 +32,22 @@ class LayoutBuilderWidget extends StatelessWidget {
       body: Center(
         child: LayoutBuilder(
           builder: (context, constraints) {
-            if(constraints.maxWidth < 600 ){
+            if (constraints.maxWidth < 600) {
               return Row(
                 mainAxisSize: MainAxisSize.max,
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: <Widget>[
-                    Expanded(
-                      child: Container(color: Colors.lightBlue),
-                      flex: 1,
-                    ),
-                    Expanded(
-                      child: Container(color: Colors.indigo),
-                      flex: 1,
-                    )
+                  Expanded(
+                    child: Container(color: Colors.lightBlue),
+                    flex: 1,
+                  ),
+                  Expanded(
+                    child: Container(color: Colors.indigo),
+                    flex: 1,
+                  )
                 ],
               );
-            }else{
+            } else {
               return Row(
                 mainAxisSize: MainAxisSize.max,
                 mainAxisAlignment: MainAxisAlignment.center,
