@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_widget_guide/utils.dart';
 import 'package:url_launcher/url_launcher.dart';
 
+import '../Ads.dart';
 import '../Code.dart';
 import '../CodeScreen.dart';
 
@@ -12,6 +13,14 @@ class RichTextWidget extends StatefulWidget {
 }
 
 class _RichTextWidgetState extends State<RichTextWidget> {
+
+  @override
+  void initState() {
+    //Hide banner ad if it isn't already hidden
+    Ads.hideBannerAd();
+    super.initState();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -44,7 +53,7 @@ class _RichTextWidgetState extends State<RichTextWidget> {
             style: TextStyle(
               fontWeight: FontWeight.normal,
               fontSize: 14.0,
-              color: Colors.black,
+              color: Theme.of(context).backgroundColor,
               height: 1.5,
             ),
             children: <TextSpan>[
