@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import '../Ads.dart';
+
 class SafeAreaWidget extends StatefulWidget {
   @override
   _SafeAreaWidgetState createState() => _SafeAreaWidgetState();
@@ -8,6 +10,13 @@ class SafeAreaWidget extends StatefulWidget {
 class _SafeAreaWidgetState extends State<SafeAreaWidget> {
   ///Bool value to control the behaviour of SafeArea widget.
   bool _isEnabled = true;
+
+  @override
+  void initState() {
+    //Hide banner ad if it isn't already hidden
+    Ads.hideBannerAd();
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -24,7 +33,7 @@ class _SafeAreaWidgetState extends State<SafeAreaWidget> {
               color: Colors.blue,
               child: Text(
                 "This widget is below safe area. If you remove the SafeArea "
-                    "widget then this text will be behind the notch.",
+                "widget then this text will be behind the notch.",
                 textAlign: TextAlign.center,
                 style: TextStyle(color: Colors.white),
               ),
@@ -35,8 +44,8 @@ class _SafeAreaWidgetState extends State<SafeAreaWidget> {
               textColor: Colors.white,
               color: Colors.indigo,
               onPressed: () => setState(() {
-                    _isEnabled == true ? _isEnabled = false : _isEnabled = true;
-                  }),
+                _isEnabled == true ? _isEnabled = false : _isEnabled = true;
+              }),
               child: Text(_isEnabled ? "Disable SafeArea" : "Enable SafeArea"),
             ),
 
