@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_widget_guide/utils.dart';
 
+import '../Ads.dart';
 import '../Code.dart';
 import '../CodeScreen.dart';
 
@@ -17,6 +18,13 @@ class _SpacerWidgetState extends State<SpacerWidget> {
   String fabText = "Remove Spacer";
 
   IconData fabIcon = Icons.cancel;
+
+  @override
+  void initState() {
+    //Hide banner ad if it isn't already hidden
+    Ads.hideBannerAd();
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -49,9 +57,12 @@ class _SpacerWidgetState extends State<SpacerWidget> {
             Padding(
               padding: const EdgeInsets.only(bottom: 18.0),
               child: Text(
-                  "Empty spaces between the boxes below are Spacer widgets"),
+                "Empty spaces between the boxes below are Spacer widgets",
+                textAlign: TextAlign.center,
+              ),
             ),
             Row(
+              mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
                 /// Empty space with flex = 2
                 isSpacerEnabled
@@ -96,7 +107,9 @@ class _SpacerWidgetState extends State<SpacerWidget> {
       floatingActionButton: FloatingActionButton.extended(
         backgroundColor: fabColor,
         onPressed: () => setState(() {
-          isSpacerEnabled == true ? isSpacerEnabled = false : isSpacerEnabled = true;
+          isSpacerEnabled == true
+              ? isSpacerEnabled = false
+              : isSpacerEnabled = true;
           isSpacerEnabled ? fabColor = Colors.red : fabColor = Colors.green;
           isSpacerEnabled ? fabIcon = Icons.cancel : fabIcon = Icons.add_circle;
           isSpacerEnabled ? fabText = "Remove Spacer" : fabText = "Add Spacer";
