@@ -51,8 +51,7 @@ class ListViewModel {
           mediumUrl: "${Utils.fadeTransitionMediumUrl}",
           videoUrl: "${Utils.fadeTransitionVideoUrl}"),
       ListItem(
-          title:
-              "${Utils.floatingActionButton.replaceAll(new RegExp(r'/'), '')}",
+          title: "${Utils.floatingActionButton.replaceAll(new RegExp(r'/'), '')}",
           description: "${Utils.floatingActionButtonDescription}",
           url: "${Utils.floatingActionButtonURL}",
           mediumUrl: "${Utils.fabMediumUrl}",
@@ -291,6 +290,216 @@ class ListViewModel {
           url: "${Utils.indexedStackURL}",
           mediumUrl: "${Utils.indexedStackMediumUrl}",
           videoUrl: "${Utils.indexedStackVideoUrl}"),
+      ListItem(
+          title: "${Utils.semantics.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.semanticsDescription}",
+          url: "${Utils.semanticsURL}",
+          mediumUrl: "${Utils.semanticsMediumURL}",
+          videoUrl: "${Utils.semanticsVideoURL}"),
+      ListItem(
+          title: "${Utils.constrainedBox.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.constrainedBoxDescription}",
+          url: "${Utils.constrainedBoxURL}",
+          mediumUrl: "${Utils.constrainedBoxMediumURL}",
+          videoUrl: "${Utils.constrainedBoxVideoURL}"),
+      ListItem(
+          title: "${Utils.stack.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.stackDescription}",
+          url: "${Utils.stackURL}",
+          mediumUrl: "${Utils.stackMediumURL}",
+          videoUrl: "${Utils.stackVideoURL}"),
+      ListItem(
+          title: "${Utils.animatedOpacity.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.animatedOpacityDescription}",
+          url: "${Utils.animatedOpacityURL}",
+          mediumUrl: "${Utils.animatedOpacityMediumURL}",
+          videoUrl: "${Utils.animatedOpacityVideoURL}"),
+      ListItem(
+          title: "${Utils.fractionallySizedBox.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.fractionallySizedBoxDescription}",
+          url: "${Utils.fractionallySizedBoxURL}",
+          mediumUrl: "${Utils.fractionallySizedBoxMediumURL}",
+          videoUrl: "${Utils.fractionallySizedBoxVideoURL}"),
+      ListItem(
+          title: "${Utils.listView.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.listViewDescription}",
+          url: "${Utils.listViewURL}",
+          mediumUrl: "${Utils.listViewMediumURL}",
+          videoUrl: "${Utils.listViewVideoURL}"),
+      ListItem(
+          title: "${Utils.listTile.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.listTileDescription}",
+          url: "${Utils.listTileURL}",
+          mediumUrl: "${Utils.listTileMediumURL}",
+          videoUrl: "${Utils.listTileVideoURL}"),
+      ListItem(
+          title: "${Utils.container.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.containerDescription}",
+          url: "${Utils.containerURL}",
+          mediumUrl: "${Utils.containerMediumURL}",
+          videoUrl: "${Utils.containerVideoURL}"),
+      ListItem(
+          title: "${Utils.selectableText.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.selectableTextDescription}",
+          url: "${Utils.selectableTextURL}",
+          mediumUrl: "${Utils.selectableTextMediumURL}",
+          videoUrl: "${Utils.selectableTextVideoURL}"),
+      ListItem(
+          title: "${Utils.dataTable.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.dataTableDescription}",
+          url: "${Utils.dataTableURL}",
+          mediumUrl: "${Utils.dataTableMediumURL}",
+          videoUrl: "${Utils.dataTableVideoURL}"),
+      ListItem(
+          title: "${Utils.slider.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.sliderDescription}",
+          url: "${Utils.sliderURL}",
+          mediumUrl: "${Utils.sliderMediumURL}",
+          videoUrl: "${Utils.sliderVideoURL}"),
+      ListItem(
+          title: "${Utils.alertDialog.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.alertDialogDescription}",
+          url: "${Utils.alertDialogURL}",
+          mediumUrl: "${Utils.alertDialogMediumURL}",
+          videoUrl: "${Utils.alertDialogVideoURL}"),
+      ListItem(
+          title: "${Utils.animatedCrossFade.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.animatedCrossFadeDescription}",
+          url: "${Utils.animatedCrossFadeURL}",
+          mediumUrl: "${Utils.animatedCrossFadeMediumURL}",
+          videoUrl: "${Utils.animatedCrossFadeVideoURL}"),
+      ListItem(
+          title: "${Utils.draggableScrollableSheet.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.draggableScrollableSheetDescription}",
+          url: "${Utils.draggableScrollableSheetURL}",
+          mediumUrl: "${Utils.draggableScrollableSheetMediumURL}",
+          videoUrl: "${Utils.draggableScrollableSheetVideoURL}"),
+      ListItem(
+          title: "${Utils.colorFiltered.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.colorFilteredDescription}",
+          url: "${Utils.colorFilteredURL}",
+          mediumUrl: "${Utils.colorFilteredMediumURL}",
+          videoUrl: "${Utils.colorFilteredVideoURL}"),
+      ListItem(
+          title: "${Utils.toggleButtons.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.toggleButtonsDescription}",
+          url: "${Utils.toggleButtonsURL}",
+          mediumUrl: "${Utils.toggleButtonsMediumURL}",
+          videoUrl: "${Utils.toggleButtonsVideoURL}"),
+      ListItem(
+          title: "${Utils.cupertinoActionSheet.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.cupertinoActionSheetDescription}",
+          url: "${Utils.cupertinoActionSheetURL}",
+          mediumUrl: "${Utils.cupertinoActionSheetMediumURL}",
+          videoUrl: "${Utils.cupertinoActionSheetVideoURL}"),
+      ListItem(
+          title: "${Utils.tweenAnimationBuilder.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.tweenAnimationBuilderDescription}",
+          url: "${Utils.tweenAnimationBuilderURL}",
+          mediumUrl: "${Utils.tweenAnimationBuilderMediumURL}",
+          videoUrl: "${Utils.tweenAnimationBuilderVideoURL}"),
+      ListItem(
+          title: "${Utils.image.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.imageDescription}",
+          url: "${Utils.imageURL}",
+          mediumUrl: "${Utils.imageMediumURL}",
+          videoUrl: "${Utils.imageVideoURL}"),
+      ListItem(
+          title: "${Utils.tabs.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.tabsDescription}",
+          url: "${Utils.tabsURL}",
+          mediumUrl: "${Utils.tabsMediumURL}",
+          videoUrl: "${Utils.tabsVideoURL}"),
+      ListItem(
+          title: "${Utils.drawer.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.drawerDescription}",
+          url: "${Utils.drawerURL}",
+          mediumUrl: "${Utils.drawerMediumURL}",
+          videoUrl: "${Utils.drawerVideoURL}"),
+      ListItem(
+          title: "${Utils.snackBar.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.snackBarDescription}",
+          url: "${Utils.snackBarURL}",
+          mediumUrl: "${Utils.snackBarMediumURL}",
+          videoUrl: "${Utils.snackBarVideoURL}"),
+      ListItem(
+          title: "${Utils.listWheelScrollView.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.listWheelScrollViewDescription}",
+          url: "${Utils.listWheelScrollViewURL}",
+          mediumUrl: "${Utils.listWheelScrollViewMediumURL}",
+          videoUrl: "${Utils.listWheelScrollViewVideoURL}"),
+      ListItem(
+          title: "${Utils.shaderMask.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.shaderMaskDescription}",
+          url: "${Utils.shaderMaskURL}",
+          mediumUrl: "${Utils.shaderMaskMediumURL}",
+          videoUrl: "${Utils.shaderMaskVideoURL}"),
+      ListItem(
+          title: "${Utils.notificationListener.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.notificationListenerDescription}",
+          url: "${Utils.notificationListenerURL}",
+          mediumUrl: "${Utils.notificationListenerMediumURL}",
+          videoUrl: "${Utils.notificationListenerVideoURL}"),
+      ListItem(
+          title: "${Utils.builder.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.builderDescription}",
+          url: "${Utils.builderURL}",
+          mediumUrl: "${Utils.builderMediumURL}",
+          videoUrl: "${Utils.builderVideoURL}"),
+      ListItem(
+          title: "${Utils.clipPath.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.clipPathDescription}",
+          url: "${Utils.clipPathURL}",
+          mediumUrl: "${Utils.clipPathMediumURL}",
+          videoUrl: "${Utils.clipPathVideoURL}"),
+      ListItem(
+          title: "${Utils.circularProgressIndicator.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.circularProgressIndicatorDescription}",
+          url: "${Utils.circularProgressIndicatorURL}",
+          mediumUrl: "${Utils.circularProgressIndicatorMediumURL}",
+          videoUrl: "${Utils.circularProgressIndicatorVideoURL}"),
+      ListItem(
+          title: "${Utils.divider.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.dividerDescription}",
+          url: "${Utils.dividerURL}",
+          mediumUrl: "${Utils.dividerMediumURL}",
+          videoUrl: "${Utils.dividerVideoURL}"),
+      ListItem(
+          title: "${Utils.ignorePointer.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.ignorePointerDescription}",
+          url: "${Utils.ignorePointerURL}",
+          mediumUrl: "${Utils.ignorePointerMediumURL}",
+          videoUrl: "${Utils.ignorePointerVideoURL}"),
+      ListItem(
+          title: "${Utils.cupertinoActivityIndicator.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.cupertinoActivityIndicatorDescription}",
+          url: "${Utils.cupertinoActivityIndicatorURL}",
+          mediumUrl: "${Utils.cupertinoActivityIndicatorMediumURL}",
+          videoUrl: "${Utils.cupertinoActivityIndicatorVideoURL}"),
+      ListItem(
+          title: "${Utils.clipOval.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.clipOvalDescription}",
+          url: "${Utils.clipOvalURL}",
+          mediumUrl: "${Utils.clipOvalMediumURL}",
+          videoUrl: "${Utils.clipOvalVideoURL}"),
+      ListItem(
+          title: "${Utils.animatedWidget.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.animatedWidgetDescription}",
+          url: "${Utils.animatedWidgetURL}",
+          mediumUrl: "${Utils.animatedWidgetMediumURL}",
+          videoUrl: "${Utils.animatedWidgetVideoURL}"),
+      ListItem(
+          title: "${Utils.padding.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.paddingDescription}",
+          url: "${Utils.paddingURL}",
+          mediumUrl: "${Utils.paddingMediumURL}",
+          videoUrl: "${Utils.paddingVideoURL}"),
+      ListItem(
+          title: "${Utils.rateApp.replaceAll(new RegExp(r'/'), '')}",
+          description: "${Utils.rateAppDescription}",
+          url: "${Utils.rateAppURL}",
+          mediumUrl: "${Utils.rateAppMediumURL}",
+          videoUrl: "${Utils.rateAppVideoURL}"),
     ];
   }
 }
